feat: base code off of sprite follow example
Prior to this the code was just an empty project that you'd get by
running `amethyst new <name>`.

This project bases the repo off of one of the new amethyst examples that
much more closely matches what our game will do ultimately and should
act as a better starting point for work.

https://github.com/amethyst/amethyst/pull/1099