feat: enable normalized orthographic camera
Prior to this change, the camera was orthographic but the view was
scaled to a fixed width and length. This causes the camera view to
stretch and distort the assets if you resize the window in a way that
changes the aspect ratio.

This change enables a builtin feature of amethyst that normalizes the
camera and automatically updates the view to match the view window in a
way that guaruntees that it maintains a fixed aspect ratio.