Add grouper system (#23)
* feat: sprite grouping proof of concept

This change adds some snap to group logic for allies. If you walk near
an allie it will snap to the next spot in the group. The number of
allies you can get is limited to 8, after which the index they think
they should go to wraps and they snap start overlapping. You can still
pick them up but it wont show independently of existing sprites.

Github issue #14