feat: Camera detached from player (#25)
Before this, the camera used the player as its Parent component, which would make the camera center on the player, but this caused some less than 100% desirable visual artifacts and made rotating the player character affect the camera.

This change decouples the player and the camera and adds a new system that updates the cameras position from the player. The rule of the update is that if the distance between the two positions becomes too great it will move the camera by the diff needed to change the distance back to the max distance. This change has the effect of causing the camera to be dragged along by the player character when it nears the edge of the screen.

This rule is probably not the best camera movement approach for a game that involves exploring new areas, but it's an acceptable first step. We can adjust it once we have other game elements in place.