Ally: added swarming (#31)
Ally movement is now defined as trying to optimize the distance to
allies and the player. The allies do not like to be close to each other
and do not want to be too far from each other or the player.

A config file was added to aid in tuning.